Support triple terms.
